Add logistics chests
Adds support for simple logistics chests. A requester chest on the outside puts down a passive provider inside and transfers items to the inside. Any other logistics chest on the outside puts down a requester chest on the inside and transfers items to the outside. The output chest will have 1 stack of each item stocked.