Updated Belt handling methods
The belts are now handled on a per lua::TransportLine basis on a timed delay based on game.tick and a stored 'next_tick' value. the increment of next_tick varies based on the outcome of how many items the line passes between the building 'walls'. if 4 items are passed, the delta on the next_tick will decrease by a predetermined amount, otherwise it will increase.