Remove unneeded inserter code.
This functionality was already commented out and is simply being removed
as it seems that the new entity type for the factory results in
inserters attempting to put goods into it regardless of the fact that it
isn't a kind that actually receives any. Nonetheless, this code should
remain in commit history in the event that it needs to be resurrected
due to game changes.